+++ src/olmo_core/distributed/checkpoint/__init__.py
@@ -171,21 +171,31 @@ def unshard_checkpoint(
     dir: PathOrStr,
     target_dir: PathOrStr,
     *,
-    optim: bool = True,
+    optim: Optional[bool] = None,
     save_overwrite: bool = False,
+    use_safetensors: bool = False,
 ) -> Tuple[Path, Optional[Path]]:
     """
     Convert a checkpoint saved via :func:`save_model_and_optim_state()` into unsharded
-    model and optimizer checkpoint files that can be loaded directly with :func:`torch.load()`.
+    model and optimizer checkpoint files that can be loaded directly with :func:`torch.load()`
+    or `safetensors <https://github.com/huggingface/safetensors>`_ if ``use_safetensors=True``.
 
     .. warning::
-        This should only be called in a non-distributed context. Otherwise a RuntimeError is raised.
+        The safetensors format cannot be used to save optimizer state, since optimizer state
+        can contain arbitrary Python objects that need to be pickled.
+        Therefore ``optim=True`` and ``use_safetensors=True`` is incompatible.
+
+    .. warning::
+        This should only be called in a non-distributed context. Otherwise a :class:`RuntimeError` is raised.
 
     :param dir: The path/URL to the original checkpoint created via :func:`save_model_and_optim_state()`.
     :param target_dir: The directory to save the unsharded model/optimizer checkpoint files to.
         This must be a local directory. URLs are not supported.
-    :param optim: Whether to unshard the optimizer state.
+    :param optim: Whether to unshard the optimizer state. This defaults to ``True`` as long as
+        ``use_safetensors=False``.
     :param save_overwrite: Overwrite any existing files in ``target_dir``.
+    :param use_safetensors: Save the unsharded files with :func:`safetensors.torch.save_file()` instead
+        of :func:`torch.save()`.
 
     :return: The path to the unsharded model checkpoint and the path to the unsharded
         optimizer checkpoint if ``optim=True``.
@@ -197,17 +207,36 @@ def unshard_checkpoint(
     from torch.distributed.checkpoint.default_planner import _EmptyStateDictLoadPlanner
     from torch.distributed.checkpoint.state_dict_loader import _load_state_dict
 
+    if optim is None:
+        optim = not use_safetensors
+    elif optim and use_safetensors:
+        raise NotImplementedError("`optim=True` is incompatible with `use_safetensors=True`")
+
     if is_distributed():
         raise RuntimeError("'unshard_checkpoint' cannot be called in a distributed context")
+
+    def save(state_dict: Dict[str, Any], path: Path):
+        if path.is_file() and not save_overwrite:
+            raise FileExistsError(
+                f"'{path}' already exists, use `save_overwrite=True` to overwrite it"
+            )
+
+        if use_safetensors:
+            from safetensors.torch import save_file
+
+            save_file(state_dict, path)
+        else:
+            torch.save(state_dict, path)
 
     dir = normalize_path(dir)
     if is_url(target_dir):
         raise ValueError("'target_dir' must be a local directory")
-    target_dir = Path(
-        _prepare_env_for_save(target_dir, save_overwrite=save_overwrite, no_dist=True)
-    )
-    model_path = target_dir / "model.pt"
-    optim_path = target_dir / "optim.pt" if optim else None
+    target_dir = Path(normalize_path(target_dir))
+    target_dir.mkdir(exist_ok=True, parents=True)
+
+    ext = "pt" if not use_safetensors else "safetensors"
+    model_path = target_dir / f"model.{ext}"
+    optim_path = target_dir / f"optim.{ext}" if optim else None
 
     model_sd: Dict[str, Any] = {}
     _load_state_dict(
@@ -218,7 +247,7 @@ def unshard_checkpoint(
     )
     if not model_sd:
         raise RuntimeError("no model state found in checkpoint")
-    torch.save(model_sd["model"], model_path)
+    save(model_sd["model"], model_path)
     del model_sd
     gc_cuda()
 
@@ -232,7 +261,7 @@ def unshard_checkpoint(
         )
         if not optim_sd:
             raise RuntimeError("no optimizer state found in checkpoint")
-        torch.save(optim_sd["optim"], optim_path)
+        save(optim_sd["optim"], optim_path)
         del optim_sd
         gc_cuda()
 
@@ -243,28 +272,25 @@ def _prepare_env_for_save(
     dir: PathOrStr,
     process_group: Optional[dist.ProcessGroup] = None,
     save_overwrite: bool = False,
-    no_dist: bool = False,
 ) -> str:
     dir = normalize_path(dir)
 
     # Prepare checkpoint folder.
     if save_overwrite:
-        if no_dist or get_fs_local_rank(process_group) == 0:
+        if get_fs_local_rank(process_group) == 0:
             clear_directory(dir)
     elif not dir_is_empty(dir):
         raise FileExistsError(dir)
 
-    if not no_dist:
-        barrier(process_group)
+    barrier(process_group)
 
     if not is_url(dir):
-        if no_dist or get_fs_local_rank(process_group) == 0:
+        if get_fs_local_rank(process_group) == 0:
             Path(dir).mkdir(exist_ok=True, parents=True)
         # Ensure the dir exists for all ranks before continuing. This might take a second if we're
         # saving to an NFS drive or something like that.
         wait_for(Path(dir).exists, description=f"Waiting on '{dir}' to be created...")
-        if not no_dist:
-            barrier(process_group)
+        barrier(process_group)
 
     return dir